Constraint-based fault monitoring for power distribution
Created a new system to insure that a value (like a current draw for example) stays between two values and create a fault if it is outside. Mainly just for monitoring the power dist. panel but could be useful elsewhere. Ask me if you're not sure how to use this. -Gabe